Presented Project - Need to add filter on stock

+++ src/PresentationModule/JLayourForms/JAdminModule.java
@@ -74,8 +74,7 @@ public class JAdminModule extends javax.swing.JFrame
      */
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents()
-    {
+    private void initComponents() {
 
         tpAdminControls = new javax.swing.JTabbedPane();
         pnlManageStaff = new javax.swing.JPanel();
@@ -111,23 +110,18 @@ public class JAdminModule extends javax.swing.JFrame
         setTitle("The Belgium Campus Stationary Management System - Administration Module");
 
         tblViewUsers.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][]
-            {
+            new Object [][] {
 
             },
-            new String []
-            {
+            new String [] {
                 "ID", "Name", "Surname", "Campus Name", "Department Name", "Cell No", "Email", "Username", "Password"
             }
-        )
-        {
-            boolean[] canEdit = new boolean []
-            {
+        ) {
+            boolean[] canEdit = new boolean [] {
                 false, false, false, false, false, false, false, false, false
             };
 
-            public boolean isCellEditable(int rowIndex, int columnIndex)
-            {
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
                 return canEdit [columnIndex];
             }
         });
@@ -135,10 +129,8 @@ public class JAdminModule extends javax.swing.JFrame
         jScrollPane2.setViewportView(tblViewUsers);
 
         btnViewUsersBack.setText("Back");
-        btnViewUsersBack.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnViewUsersBack.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnViewUsersBackMouseClicked(evt);
             }
         });
@@ -169,23 +161,18 @@ public class JAdminModule extends javax.swing.JFrame
         tpAdminControls.addTab("Manage Staff", pnlManageStaff);
 
         tblRegisterRequests.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][]
-            {
+            new Object [][] {
 
             },
-            new String []
-            {
+            new String [] {
                 "Request Date", "Name", "Surname", "Campus Name", "Department Name", "Cell No", "Email", "Username", "Password"
             }
-        )
-        {
-            boolean[] canEdit = new boolean []
-            {
+        ) {
+            boolean[] canEdit = new boolean [] {
                 false, false, false, false, false, false, false, false, false
             };
 
-            public boolean isCellEditable(int rowIndex, int columnIndex)
-            {
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
                 return canEdit [columnIndex];
             }
         });
@@ -193,28 +180,22 @@ public class JAdminModule extends javax.swing.JFrame
         jScrollPane1.setViewportView(tblRegisterRequests);
 
         btnRegisterUserBack.setText("Back");
-        btnRegisterUserBack.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnRegisterUserBack.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnRegisterUserBackMouseClicked(evt);
             }
         });
 
         btnDeleteRegisterUser.setText("Delete User");
-        btnDeleteRegisterUser.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnDeleteRegisterUser.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnDeleteRegisterUserMouseClicked(evt);
             }
         });
 
         btnAcceptUser.setText("Accept User");
-        btnAcceptUser.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnAcceptUser.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnAcceptUserMouseClicked(evt);
             }
         });
@@ -252,52 +233,41 @@ public class JAdminModule extends javax.swing.JFrame
         pnlUserRegistrationRequests.getAccessibleContext().setAccessibleName("");
 
         tblAllStock.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][]
-            {
+            new Object [][] {
 
             },
-            new String []
-            {
+            new String [] {
                 "Stationary Stock ID", "Product Name", "Category Name", "Model", "Price", "Quantity", "Date Of Entry / Update"
             }
-        )
-        {
-            boolean[] canEdit = new boolean []
-            {
+        ) {
+            boolean[] canEdit = new boolean [] {
                 false, false, false, false, false, false, false
             };
 
-            public boolean isCellEditable(int rowIndex, int columnIndex)
-            {
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
                 return canEdit [columnIndex];
             }
         });
         jScrollPane3.setViewportView(tblAllStock);
 
         btnViewAllStockBack.setLabel("Back");
-        btnViewAllStockBack.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnViewAllStockBack.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnViewAllStockBackMouseClicked(evt);
             }
         });
 
         btnPromptUpdate.setText("Update");
         btnPromptUpdate.setToolTipText("");
-        btnPromptUpdate.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnPromptUpdate.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnPromptUpdateMouseClicked(evt);
             }
         });
 
         btnDeleteStock.setText("Delete");
-        btnDeleteStock.addMouseListener(new java.awt.event.MouseAdapter()
-        {
-            public void mouseClicked(java.awt.event.MouseEvent evt)
-            {
+        btnDeleteStock.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
                 btnDeleteStockMouseClicked(evt);
             }
         });
@@ -336,10 +306,8 @@ public class JAdminModule extends javax.swing.JFrame
         jmAdminApplication.setText("Application");
 
         jmAdminLogout.setText("Logout");
-        jmAdminLogout.addActionListener(new java.awt.event.ActionListener()
-        {
-            public void actionPerformed(java.awt.event.ActionEvent evt)
-            {
+        jmAdminLogout.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jmAdminLogoutActionPerformed(evt);
             }
         });
@@ -347,10 +315,8 @@ public class JAdminModule extends javax.swing.JFrame
         jmAdminApplication.add(jSeparator1);
 
         jmAdminClose.setText("Close");
-        jmAdminClose.addActionListener(new java.awt.event.ActionListener()
-        {
-            public void actionPerformed(java.awt.event.ActionEvent evt)
-            {
+        jmAdminClose.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jmAdminCloseActionPerformed(evt);
             }
         });
@@ -361,20 +327,16 @@ public class JAdminModule extends javax.swing.JFrame
         jmAccounts.setText("Accounts");
 
         jmManageRegisteredUsers.setText("Manage Registered Users");
-        jmManageRegisteredUsers.addActionListener(new java.awt.event.ActionListener()
-        {
-            public void actionPerformed(java.awt.event.ActionEvent evt)
-            {
+        jmManageRegisteredUsers.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jmManageRegisteredUsersActionPerformed(evt);
             }
         });
         jmAccounts.add(jmManageRegisteredUsers);
 
         jmViewRegistrationRequests.setText("View Registration Requests");
-        jmViewRegistrationRequests.addActionListener(new java.awt.event.ActionListener()
-        {
-            public void actionPerformed(java.awt.event.ActionEvent evt)
-            {
+        jmViewRegistrationRequests.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jmViewRegistrationRequestsActionPerformed(evt);
             }
         });
@@ -388,10 +350,8 @@ public class JAdminModule extends javax.swing.JFrame
 
         jmManageAllStock.setText("Manage All Stock");
         jmManageAllStock.setActionCommand("");
-        jmManageAllStock.addActionListener(new java.awt.event.ActionListener()
-        {
-            public void actionPerformed(java.awt.event.ActionEvent evt)
-            {
+        jmManageAllStock.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jmManageAllStockActionPerformed(evt);
             }
         });
@@ -399,10 +359,8 @@ public class JAdminModule extends javax.swing.JFrame
 
         jmInserStock.setText("Insert Stock");
         jmInserStock.setActionCommand("Insert New Stock");
-        jmInserStock.addActionListener(new java.awt.event.ActionListener()
-        {
-            public void actionPerformed(java.awt.event.ActionEvent evt)
-            {
+        jmInserStock.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jmInserStockActionPerformed(evt);
             }
         });
@@ -533,7 +491,7 @@ public class JAdminModule extends javax.swing.JFrame
             Helper.DisplayError(aiob.toString());
         } catch (Exception ex)
         {
-            Helper.DisplayError(ex.getMessage().toString());
+            Helper.DisplayError(ex.getMessage());
         }
     }
 
@@ -717,11 +675,16 @@ public class JAdminModule extends javax.swing.JFrame
     private void btnAcceptUserMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnAcceptUserMouseClicked
         try
         {
-            selectedStaff.InsertStaff();
-            registerStaff.remove(selectedStaff);
-            tblRegisterRequests.setModel(new DefaultTableModel());
-            PopulateTableRegister();
-            JOptionPane.showConfirmDialog(this, "User succesfully Accepted", "Registration Completed", JOptionPane.DEFAULT_OPTION);
+            int choice = JOptionPane.showConfirmDialog(this, "Are you sure you want to accept this user registration?", "Confirmation", JOptionPane.YES_NO_OPTION);
+
+            if (choice == JOptionPane.YES_OPTION)
+            {
+                selectedStaff.InsertStaff();
+                registerStaff.remove(selectedStaff);
+                tblRegisterRequests.setModel(new DefaultTableModel());
+                PopulateTableRegister();
+                JOptionPane.showConfirmDialog(this, "User succesfully Accepted", "Registration Completed", JOptionPane.DEFAULT_OPTION);
+            }
         } catch (NullPointerException npe)
         {
             Helper.DisplayError("There is no data currently available in the table");
@@ -731,15 +694,20 @@ public class JAdminModule extends javax.swing.JFrame
     private void btnDeleteRegisterUserMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnDeleteRegisterUserMouseClicked
         try
         {
-            SelectRowRegister();
-            selectedStaff.DeleteRegistry();
-            registerStaff.remove(selectedStaff);
+            int choice = JOptionPane.showConfirmDialog(this, "Are you sure you want to delete this user registration??", "Confirmation", JOptionPane.YES_NO_OPTION);
 
-            DefaultTableModel model = (DefaultTableModel) tblRegisterRequests.getModel();
-            model.setRowCount(0);
+            if (choice == JOptionPane.YES_OPTION)
+            {
+                SelectRowRegister();
+                selectedStaff.DeleteRegistry();
+                registerStaff.remove(selectedStaff);
 
-            PopulateTableRegister();
-            JOptionPane.showConfirmDialog(this, "User succesfully Removed", "Registration Completed", JOptionPane.DEFAULT_OPTION);
+                DefaultTableModel model = (DefaultTableModel) tblRegisterRequests.getModel();
+                model.setRowCount(0);
+
+                PopulateTableRegister();
+                JOptionPane.showConfirmDialog(this, "User succesfully Removed", "Registration Completed", JOptionPane.DEFAULT_OPTION);
+            }
         } catch (NullPointerException npe)
         {
             Helper.DisplayError("There is no data currently available in the table");
